when converting document node, assoc doctype as metadata on root element
I would like to preserve the information XOM places in the Document
node, but at the same time I'd like to maintain compatability with
clojure.xml conventions. Instead of wrapping the root element in some
kind of {:document-type ... :root-element ...} map, we'll associate
the important information from the document (it's doc-type directly
onto the root element's metadata map. This means that a root
element (representing a document) is distinguishable from a any other
element only by virtue of the fact that it has a :doc-type key in its
metadata map.